+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## 2.8.0 - 2019-11-08
 
 * Added: Cmdlets to get and modify certificate private key permission
 * Changed: Return the certificate object in New-DomainSignedCertificate